// src/types/classType.ts
export type ClassType = {
  SEC: string;
  CLASS: string;
  NOMATERIALCOST: string;
  RESERVECAPACITY: string;
  CLASSNOTES: string;
  TYPE: string;
  DAYS: string;
  TIME: string;
  LOCATION: string;
  INSTRUCTOR: string;
  COMMENT: string;
  COURSECODE: string;
  COURSETITLE: string;
  Units: string;
  id?: number;
};